Removed Cargo.lock
We don't want Cargo.lock to be present in this template, as this template is not getting updated on every released Bevy patch, yet we want projects that use this template to use the very most up-to-date version available. I haven't added Cargo.lock to .gitignore, as that would also be reflected on projects using this template, which is not desirable.